Silence some older warnings
`-Wdeclaration-after-statement` was previously forced on because back in
the days of C90, that was a hard requirement. It isn't anymore, and
allowing inline declarations (without having to scope change) is just
flat out more convenient, and allows for better organization. It also
removes warnings when trying to apply that flag to C++ files.